Use Path.Combine() for concatenating paths instead of "+"

+++ Queue/QueueData.cs
@@ -380,7 +380,7 @@ namespace MatterHackers.MatterControl.PrintQueue
 
 			}
 
-			if (!File.Exists(locationToSaveTo))
+			if (!File.Exists(locationToSaveTo) && File.Exists(fileNameToLoad))
 			{
 				File.Copy(fileNameToLoad, locationToSaveTo);
 			}
